refactor: move wallet utility functions into useWallet hook for better organization
