--- src/utils/hook/wallet.hook.ts
@@ -1,25 +1,0 @@
-import { useWallet } from "@/wallet/hooks/useWallet.hook";
-
-export const useWalletUtils = () => {
-  const { connectWallet, disconnectWallet } = useWallet();
-
-  const handleConnect = async () => {
-    try {
-      await connectWallet();
-    } catch (error) {
-      console.error("Error connecting wallet:", error);
-    }
-  };
-
-  const handleDisconnect = async () => {
-    try {
-      if (disconnectWallet) {
-        await disconnectWallet();
-      }
-    } catch (error) {
-      console.error("Error disconnecting wallet:", error);
-    }
-  };
-
-  return { handleConnect, handleDisconnect };
-};
